ElExplanation: move to "results" section
