+++ ui-components/src/components/results/ElExplanation.vue
